db connection

+++ src/controller/UserController.ts
@@ -1,17 +1,17 @@
-import { getRepository } from "typeorm"
+// import { getRepository } from "typeorm"
 import { NextFunction, Request, Response } from "express"
 import { User } from "../entity/User"
-
+import { AppDataSource } from "../data-source"
 export class UserController {
 
-    private userRepository = getRepository(User)
+    private userRepository = AppDataSource.getRepository(User)
 
     async all(request: Request, response: Response, next: NextFunction) {
         return this.userRepository.find()
     }
 
     async one(request: Request, response: Response, next: NextFunction) {
-        return this.userRepository.findOne(request.params.id)
+        return this.userRepository.findOne({where: { id: parseInt(request.params.id) }})
     }
 
     async save(request: Request, response: Response, next: NextFunction) {
@@ -19,8 +19,10 @@ export class UserController {
     }
 
     async remove(request: Request, response: Response, next: NextFunction) {
-        let userToRemove = await this.userRepository.findOneBy({ id: request.params.id })
-        await this.userRepository.remove(userToRemove)
+        let userToRemove = await this.userRepository.findOneBy({ id: parseInt(request.params.id) })
+        if (userToRemove) {
+            await this.userRepository.remove(userToRemove)
+        }
     }
 
 }
